Modified indexing on large textures to be long iso int.

+++ src/geomviewer.c
@@ -7287,7 +7287,7 @@ void _s2priv_ss2pt(unsigned int itextureID, int usemipmaps) {
       if (_s2_devcap & _S2DEVCAP_NOCOLOR) {
 	// desaturate the bitmap data
 	int i, j, k;
-	int idx = 0;
+	long idx = 0;
 	float sum;
 	for (i = 0; i < width; i++) {
 	  for (j = 0; j < height; j++) {
@@ -10575,7 +10575,7 @@ unsigned int _s2priv_setupTexture(int width, int height,
   if (_s2_devcap & _S2DEVCAP_NOCOLOR) {
     // desaturate the bitmap data
     int i, j;
-    int idx = 0;
+    long idx = 0;
     float sum;
     for (i = 0; i < width; i++) {
       for (j = 0; j < height; j++, idx++) {
@@ -10630,7 +10630,7 @@ unsigned int _s2priv_setupTexture3d(int width, int height, int depth,
   if (_s2_devcap & _S2DEVCAP_NOCOLOR) {
     // desaturate the bitmap data
     int i, j, k;
-    int idx = 0;
+    long idx = 0;
     float sum;
     for (i = 0; i < width; i++) {
       for (j = 0; j < height; j++) {
